+++ sass/styles.scss
@@ -48,18 +48,20 @@ h5 {
     margin-top: 64px;
     margin-bottom: 48px;
   }
-  h1 {
-    flex: 0 1 auto;
-    font-size: 56px;
-    text-align: center;
-    font-weight: bold;
-    color: $text-color;
-    width: 100%;
-    @include respond-to(medium) {
-      font-size: 72px;
-      width: 45%
+  a#homeLink {
+      h1 {
+        flex: 0 1 auto;
+        font-size: 56px;
+        text-align: center;
+        font-weight: bold;
+        color: $text-color;
+        width: 100%;
+        @include respond-to(medium) {
+          font-size: 72px;
+          width: 100%
+        }
+      }
     }
-  }
 }
 
 #headerDesc {
