+++ views/layout.pug
@@ -6,7 +6,7 @@ html(lang="en")
     title _ARCHITEKT
   body
     div#header
-      a(href='/')
+      a(href='/' id="homeLink")
         h1 _ARCHITEKT
       div#headerDesc Welcome to our online portfolio. We are a small group of passionate designers and architects, looking to change and create amazing digital images to inspire others to follow in our innovative steps.
     block body
